requirements.txt
fastapi
uvicorn
celery[redis]
redis
python-dotenv
requests
langchain
openai
pytest